requirements.txt
Flask==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.28
requests==2.28.1
Werkzeug==2.0.3
python-dotenv==0.20.0
stripe==12.0.0 